+++ src/Marten/Exceptions/EventStreamUnexpectedMaxEventIdException.cs
@@ -7,7 +7,7 @@ namespace Marten.Exceptions
     {
         public Type AggregateType { get; }
 
-        public EventStreamUnexpectedMaxEventIdException(object id, Type aggregateType, int expected, int actual) : base($"Unexpected MAX(id) for event stream, expected {expected} but got {actual}", aggregateType, id)
+        public EventStreamUnexpectedMaxEventIdException(object id, Type aggregateType, long expected, long actual) : base($"Unexpected MAX(id) for event stream, expected {expected} but got {actual}", aggregateType, id)
         {
             Id = id;
             AggregateType = aggregateType;
